Homework 7 & 8 (#6)
* feat: add repositories and interfaces, extend IServiceCollection with AddRepositories()

* fix: change GenericRepository interface to smaller interfaces for better SRP

* feat: add nested objects to models

* feat: comments repository methods implementation

* feat: implement sql for comments, permissions, priorities, projects, projectrolepermissions

* feat: implement CRUD for half of repositories

* feat: implemented DAL part2

* feat: implement dal part3

* feat: password hash, implement all crud methods, code refactor

* feat: implement DAL methods needed for system

* feat: add services and register them for DI

* feat: inject repositories to services

* feat: add cqrs + mediatr method

* feat: add ps script to create CQRS + handlers automatically

* feat: add powershell script for creating cqrs automatically

* feat: add CQRS and handler files

* refactor: add empty line after namespaces

* refactor: change separate repository interfaces to one generic repository interface

* fix: change the tabs in sql methods, fix few issues in sql syntax

* refactor: tweak tabulation in CommentRepository GetByUserAndProjectId method

* chore: change Hasher to static class

* chore: move Services DI registration to Application Layer

* feat: Comments flow

* feat: add request and response dtos, update controllers, add implementation to services

* feat: add transactions, add fluent validation

* chore: remove cqrs files, change DTOs classes to records, and fluentvalidators, apply pattern matching

* refactor: change DTO models to be records

* feat: add FluentValidation to all Create/Update DTO models

* refactor: separate private members with empty line above, separate class and namespace with empty line, rename IGenericRepository to IRepository

* refactor: change controller method names

* feat: add SoftDelete endpoint logic for Projects, Users, Tickets controllers

* feat: add ticket deleting transactions

* fix: remove dbConnectionFactory and replace it with IDbConnection (injecting as scoped is enough with well structured requests)

* fix: delete unnecessary fields from domain models

* feat: transactions on creating

* feat: add TicketsByProjectId, TicketsByKeyAndValue, UsersByProjectId, TagsByProjectId endpoints

* feat: add joins to get methods for Comment, ProjectRole, TicketRelation, User entities

* feat: add joins to ticket gets

* refactor: change specific types to var, change lists to IEnumerable where it was possible, add empty lines above and under constructors, multiline constructor parameters

* fix: open connection for transactions, feat: add implementation to tickethub

* feat: add DbTransactionHelper

* refactor: extract shared ticket deletion logic to private method

* refactor: change private members nomenclature to underscore nomenclature

* feat: add middleware to handle exceptions globally

* fix: mapping ticket to ticketDto

* fix: TicketNumber assignment

* fix: add "/soft-delete" at the end of the soft delte patch methods

* fix: notify ticketshub on ticket update

* fix: FluentValidator is now automatically validating DTOs

* feat: add jwt generator, authorization handler, const permissions (initial implementation of auth)

* feat: add fully working jwt token, authorize and policy handler

* refactor: add empty lines

* fix: delete "id/" from GetTicket endpoint, modify the getByKeyAndNumber endpoint

* feat: add authorize attributes to controllers and methods

* feat: add GetProjectsByUserId endpoint

* chore: cleanup Program.cs

* chore: document controllers

* feat: add projectId and adjust Policy related endpoints for ProjectPermissionHandler

* feat: add projectId Validation on CommentService

* refactor: export throw exception to private method

* feat: add projectId validation for all entities related with projects

* chore: move ProjectAccessValidator to AddAuthServices

* fix: remove route from GetProjects

* feat: add jwt to swagger, cleanup Program.cs (add extension class in API layer)

* feat: add health checks

* feat: register logger and cutsom exceptions

* refactor: change naming to projectAccessValidator

* refactor: tweak repository methods nomenclature

* feat: add logging to comments

* feat: change create methods to return booleans - guids are generated in services

* feat: add exception handling and logging to comments

* feat: add exceptions and logging to permissions

* feat: add logging and exception handling to all services

* fix: add RoleId to UserRoleCreateDto

* feat: register custom exceptions in middleware

* feat: add logging and exception for auth

* fix: User CreatedAt field not set, Permission canno be viewed, ProjectRoleCreateDto has ProjectRoleId, Key keyword not in squarebrackets for Project Update, Ticket can reffer itself in TicketRelation

* fix: Tickets can reffer between different Projects, Ticket GetById, GetAll, GetByProjectAndKey nulled relations, Get Ticket Key in sql not in brackets

* fix: projectkey can have lower case letters

* fix: deleting projects not working

* fix: exceptions thrown in cascade deletions when no entities related in the tables

* fix: projects not being able to delete

* fix: route in controllers

* feat: add SignalR token authorization

* feat: add ProjectService unit tests

* feat: add logger to Program.cs

* Update README.md

* fix: add Arrange, Act, Assert comments to tests

* fix: remove Console.WriteLine from ProjectPermissionHandler

* fix: change the method name to GetByIdWithRolesAndPermissions

* feat: add UserRegisterDto, move the Auth related Dtos from DTOs/User to DTOs/Auth

* feat: add registering to AuthController

* fix: change ProjectPemissions in UserClaimsDto to capital letter

* fix: get userId from httpContextAccessor  for "CreatedBy" entities

* fix: remove user validation from ticket update

+++ ProjectManager_01.WebAPI.Application/Services/UserService.cs
@@ -1,9 +1,11 @@
 ﻿using System.Data;
 using AutoMapper;
+using Microsoft.Extensions.Logging;
 using ProjectManager_01.Application.Contracts.Repositories;
 using ProjectManager_01.Application.Contracts.Services;
 using ProjectManager_01.Application.DTOs.UserRoles;
 using ProjectManager_01.Application.DTOs.Users;
+using ProjectManager_01.Application.Exceptions;
 using ProjectManager_01.Application.Helpers;
 using ProjectManager_01.Domain.Models;
 
@@ -18,6 +20,7 @@ public class UserService : IUserService
     private readonly IDbConnection _dbConnection;
     private readonly ICommentService _commentService;
     private readonly ITicketService _ticketService;
+    private readonly ILogger<UserService> _logger;
 
     public UserService(
         IUserRepository userRepository,
@@ -26,7 +29,8 @@ public class UserService : IUserService
         IProjectUserRoleService projectUserRoleService,
         IDbConnection dbConnection,
         ICommentService commentService,
-        ITicketService ticketService)
+        ITicketService ticketService,
+        ILogger<UserService> logger)
     {
         _userRepository = userRepository;
         _mapper = mapper;
@@ -35,43 +39,74 @@ public class UserService : IUserService
         _dbConnection = dbConnection;
         _commentService = commentService;
         _ticketService = ticketService;
+        _logger = logger;
     }
 
     public async Task CreateUserAsync(UserCreateDto userCreateDto)
     {
+        _logger.LogWarning("Creating User called. User: {UserName}", userCreateDto.UserName);
+
         using var transaction = DbTransactionHelper.BeginTransaction(_dbConnection);
 
         try
         {
             var user = _mapper.Map<User>(userCreateDto);
             user.PasswordHash = BcryptPasswordHasher.HashPassword(userCreateDto.Password);
-            var userId = await _userRepository.CreateAsync(user, transaction);
+            user.Id = Guid.NewGuid();
+            user.CreatedAt = DateTimeOffset.UtcNow;
 
-            var userRoleCreateDto = new UserRoleCreateDto(userId);
-            await _userRoleService.CreateUserRoleAsync(userRoleCreateDto, transaction);
+            // Check if operation is successful
+            if (!await _userRepository.CreateAsync(user, transaction))
+            {
+                _logger.LogError("Creating User failed. User: {UserName}", userCreateDto.UserName);
+                throw new OperationFailedException("Creating User failed.");
+            }
+
+            await _userRoleService.CreateDefaultUserRoleAsync(user.Id, transaction);
 
             transaction.Commit();
+
+            _logger.LogInformation("Creating User successful. User: {UserId}", user.Id);
         }
         catch
         {
             transaction.Rollback();
-            throw new Exception("Error while performing user creation transaction.");
+            _logger.LogError("Creating User transaction failed. User: {UserName}", userCreateDto.UserName);
+            throw;
         }
     }
 
     public async Task UpdateUserAsync(UserUpdateDto userUpdateDto)
     {
+        _logger.LogInformation("Updating User called. User: {UserId}", userUpdateDto.Id);
+
         var user = _mapper.Map<User>(userUpdateDto);
-        await _userRepository.UpdateAsync(user);
+
+        // Check if operation is successful
+        if (!await _userRepository.UpdateAsync(user))
+        {
+            _logger.LogError("Updating User failed. User: {UserId}", user.Id);
+            throw new OperationFailedException("Updating User failed.");
+        }
+
+        _logger.LogInformation("Updating User successful. User: {UserId}", user.Id);
     }
 
     public async Task DeleteUserAsync(Guid userId)
     {
+        _logger.LogWarning("Deleting User transaction called. User: {UserId}", userId);
+
         using var transaction = DbTransactionHelper.BeginTransaction(_dbConnection);
 
         try
         {
-            await _userRepository.DeleteAsync(userId, transaction);
+            // Check if operation is successful
+            if (!await _userRepository.DeleteByIdAsync(userId, transaction))
+            {
+                _logger.LogError("Deleting User failed. User: {UserId}", userId);
+                throw new OperationFailedException("Deleting User transaction failed.");
+            }
+
             await _userRoleService.DeleteByUserIdAsync(userId, transaction);
             await _projectUserRoleService.DeleteByUserIdAsync(userId, transaction);
             await _commentService.DeleteByUserIdAsync(userId, transaction);
@@ -79,37 +114,64 @@ public class UserService : IUserService
             await _ticketService.DeleteTicketByUserIdAsync(userId, transaction);
 
             transaction.Commit();
+            _logger.LogInformation("Deleting User transaction successful. User: {UserId}", userId);
         }
         catch
         {
             transaction.Rollback();
-            throw new Exception("Error while performing user deletion transaction.");
+            _logger.LogError("Deleting User transaction failed. User: {UserId}", userId);
+            throw;
         }
     }
 
     public async Task<UserDto> GetUserByIdAsync(Guid userId)
     {
+        _logger.LogInformation("Getting User called. User: {UserId}", userId);
+
         var user = await _userRepository.GetByIdAsync(userId);
 
+        // Check if operation is successful
+        if (user == null)
+        {
+            _logger.LogError("Getting User failed. User: {UserId}", userId);
+            throw new NotFoundException("User not found.");
+        }
+
+        _logger.LogInformation("Getting User successful. User: {UserId}", userId);
         return _mapper.Map<UserDto>(user);
     }
 
     public async Task<IEnumerable<UserDto>> GetAllUsersAsync()
     {
+        _logger.LogInformation("Getting all Users called.");
+
         var users = await _userRepository.GetAllAsync();
 
+        _logger.LogInformation("Getting all Users ({Count}) successful.", users.Count());
         return _mapper.Map<IEnumerable<UserDto>>(users);
     }
 
     public async Task SoftDeleteUserAsync(Guid userId)
     {
-        await _userRepository.SoftDeleteAsync(userId);
+        _logger.LogWarning("Soft deleting User called. User: {UserId}", userId);
+
+        // Check if operation is successful
+        if (!await _userRepository.SoftDeleteByIdAsync(userId))
+        {
+            _logger.LogError("Soft deleting User failed. User: {UserId}", userId);
+            throw new OperationFailedException("Soft deleting User failed.");
+        }
+
+        _logger.LogInformation("Soft deleting User successful. User: {UserId}", userId);
     }
 
     public async Task<IEnumerable<UserDto>> GetUsersByProjectIdAsync(Guid projectId)
     {
-        var users = await _userRepository.GetByProjectIdAsync(projectId);
+        _logger.LogInformation("Getting Users by ProjectId called. Project: {ProjectId}", projectId);
 
+        var users = await _userRepository.GetAllByProjectIdAsync(projectId);
+
+        _logger.LogInformation("Getting Users by ProjectId ({Count}) successful. Project: {ProjectId}", users.Count(), projectId);
         return _mapper.Map<IEnumerable<UserDto>>(users);
     }
 }
